requirements.txt
django==4.1.5
huey==2.4.4
redis==4.4.2
psycopg2-binary==2.9.5
requests==2.28.2
whitenoise==6.3.0
uuid==1.30
gunicorn==20.1.0
django-storages==1.13.2
boto3==1.26.56
django-environ==0.9.0


